+++ src/Neighbor.Contract/DTOs/MemberDTOs/LessorDTO.cs
@@ -7,6 +7,6 @@ public sealed class LessorDTO
     public Guid? AccountId { get; set; }
     public string? WareHouseAddress { get; set; }
     public string? ShopName { get; set; }
-    public int? Wallet { get; set; }
+    public long? Wallet { get; set; }
     public LocationType? LocationType { get; set; }
 }
