+++ src/Neighbor.Contract/Services/Products/Command.cs
@@ -6,7 +6,7 @@ using Neighbor.Contract.Enumarations.Product;
 namespace Neighbor.Contract.Services.Products;
 public static class Command
 {
-    public record CreateProductCommand(string Name, string Description, int Value, int Price, int MaximumRentDays, string Policies, int CategoryId, Guid UserId, List<IFormFile> ProductImages, InsuranceDTO.InsuranceRequestDTO? Insurance, List<SurchargeDTO.SurchargeRequestDTO>? ListSurcharges) : ICommand;
+    public record CreateProductCommand(string Name, string Description, long Value, long Price, int MaximumRentDays, string Policies, int CategoryId, Guid UserId, List<IFormFile> ProductImages, InsuranceDTO.InsuranceRequestDTO? Insurance, List<SurchargeDTO.SurchargeRequestDTO>? ListSurcharges) : ICommand;
 
     public record ConfirmProductCommand(Guid ProductId, ConfirmStatus ConfirmStatus, string? RejectReason) : ICommand;
