+++ src/Neighbor.Domain/Entities/Transaction.cs
@@ -5,7 +5,7 @@ namespace Neighbor.Domain.Entities;
 
 public class Transaction : DomainEntity<Guid>
 {
-    public Transaction(int amount, TransactionType type, string description, Guid walletId)
+    public Transaction(long amount, TransactionType type, string description, Guid walletId)
     {
         Amount = amount;
         Type = type;
@@ -13,18 +13,18 @@ public class Transaction : DomainEntity<Guid>
         WalletId = walletId;
     }
 
-    public int Amount { get; private set; }
+    public long Amount { get; private set; }
     public TransactionType Type { get; private set; }
     public string Description { get; private set; }
     public Guid WalletId { get; private set; }
     public virtual Wallet Wallet { get; private set; }
 
-    public static Transaction CreateTransactionWithTypeDeposit(int amount, string description, Guid walletId)
+    public static Transaction CreateTransactionWithTypeDeposit(long amount, string description, Guid walletId)
     {
         return new Transaction(amount, TransactionType.Deposit, description, walletId);
     }
 
-    public static Transaction CreateTransactionWithTypeRefund(int amount, string description, Guid walletId)
+    public static Transaction CreateTransactionWithTypeRefund(long amount, string description, Guid walletId)
     {
         return new Transaction(amount, TransactionType.Refund, description, walletId);
     }
